Refactoring Runtime.Api - ExecutionResult
Moving `ExecutionResult` out of the `Runtime` class monster.
